fix function call + func_800941C4

+++ include/functions.h
@@ -664,7 +664,7 @@ void func_8001C96C(void);
 s32 func_8001E96C(void);
 s32 func_8001B930(s32, s32);            /* extern */
 void func_8001BC84(s32, s32, s32);      /* extern */
-void func_800941C4(s16, f32, f32, f32); /* extern */
+void func_800941C4(s32, f32, f32, f32); /* extern */
 void func_8006F570(s32 arg0, s32 arg1, s32 arg2, f32 arg3, f32 arg4);
 int Decode(u8* buf_in, u8* buf_out);
 void func_8001ABF4(s32 arg0, s32 arg1, s32 arg2, void* arg3);
